Reformat imports and functions

+++ filters/endisland.h
@@ -1,3 +1,3 @@
-#include "../submodules/cubiomes/biomenoise.h"
+#include <stdint.h>
 
-int end_cage(int lower48, int mc_version, int end_cage_limit);
+int end_cage(int64_t lower48, int mc_version, int end_cage_limit);
